# fonts/bk_fonts/gateway/src/api/index.ts
/**
 * @fileoverview API Gateway para el metaverso
 * @module @metaverso/gateway/api
 */

import { Router } from 'express';
import { GatewayConfig, HealthStatus } from '../types';
import { Logger } from '../utils/logger';

/**
 * API Gateway principal
 */
export class APIGateway {
  private router: Router;
  private config: GatewayConfig;
  private logger: Logger;
  private initialized: boolean = false;

  constructor(config: GatewayConfig) {
    this.config = config;
    this.router = Router();
    this.logger = new Logger(config.monitoring.logging);
  }

  /**
   * Inicializa el API Gateway
   */
  async initialize(): Promise<void> {
    if (this.initialized) {
      return;
    }

    try {
      this.logger.info('Inicializando API Gateway...');

      // Configurar rutas de la API
      this.setupRoutes();

      this.initialized = true;
      this.logger.info('API Gateway inicializado correctamente');

    } catch (error) {
      this.logger.error('Error al inicializar API Gateway:', error);
      throw error;
    }
  }

  /**
   * Obtiene el router de Express
   */
  getRouter(): Router {
    return this.router;
  }

  /**
   * Obtiene el estado de salud
   */
  async getHealthStatus(): Promise<HealthStatus> {
    return {
      name: 'api-gateway',
      status: this.initialized ? 'healthy' : 'unhealthy',
      timestamp: new Date().toISOString(),
      details: {
        routes: this.router.stack.length,
        initialized: this.initialized
      }
    };
  }

  /**
   * Destruye el API Gateway
   */
  async destroy(): Promise<void> {
    this.initialized = false;
    this.logger.info('API Gateway destruido');
  }

  // Métodos privados

  private setupRoutes(): void {
    // Worlds API
    this.router.get('/worlds', this.getWorlds.bind(this));
    this.router.post('/worlds', this.createWorld.bind(this));
    this.router.get('/worlds/:id', this.getWorld.bind(this));

    // Users API
    this.router.get('/users/:id', this.getUser.bind(this));

    // Sessions API
    this.router.post('/sessions', this.createSession.bind(this));

    // Economy API
    this.router.get('/economy/transactions', this.getTransactions.bind(this));

    // Assets API
    this.router.get('/assets', this.getAssets.bind(this));
  }

  // Handlers de rutas

  private async getWorlds(req: any, res: any): Promise<void> {
    try {
      const { page = 1, limit = 20, category } = req.query;

      // Simular respuesta de mundos
      const worlds = [
        {
          id: '1',
          name: 'Mundo Virtual Gaming',
          description: 'Un mundo increíble para gaming',
          owner: 'user-1',
          status: 'published',
          category: 'gaming',
          capacity: 100,
          currentUsers: 25,
          createdAt: new Date().toISOString(),
          updatedAt: new Date().toISOString()
        }
      ];

      res.json({
        success: true,
        data: {
          worlds,
          pagination: {
            page: parseInt(page),
            limit: parseInt(limit),
            total: worlds.length,
            totalPages: 1
          }
        },
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al obtener mundos:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al obtener mundos',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async createWorld(req: any, res: any): Promise<void> {
    try {
      const { name, description, category, capacity = 100 } = req.body;

      // Validar datos requeridos
      if (!name || !description || !category) {
        return res.status(400).json({
          success: false,
          error: {
            code: 'BAD_REQUEST',
            message: 'Nombre, descripción y categoría son requeridos',
            timestamp: new Date().toISOString()
          }
        });
      }

      // Simular creación de mundo
      const world = {
        id: Date.now().toString(),
        name,
        description,
        owner: req.user?.id || 'anonymous',
        status: 'draft',
        category,
        capacity,
        currentUsers: 0,
        createdAt: new Date().toISOString(),
        updatedAt: new Date().toISOString()
      };

      res.status(201).json({
        success: true,
        data: world,
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al crear mundo:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al crear mundo',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async getWorld(req: any, res: any): Promise<void> {
    try {
      const { id } = req.params;

      // Simular búsqueda de mundo
      const world = {
        id,
        name: 'Mundo Virtual Gaming',
        description: 'Un mundo increíble para gaming',
        owner: 'user-1',
        status: 'published',
        category: 'gaming',
        capacity: 100,
        currentUsers: 25,
        createdAt: new Date().toISOString(),
        updatedAt: new Date().toISOString()
      };

      res.json({
        success: true,
        data: world,
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al obtener mundo:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al obtener mundo',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async getUser(req: any, res: any): Promise<void> {
    try {
      const { id } = req.params;

      // Simular búsqueda de usuario
      const user = {
        id,
        username: 'usuario_ejemplo',
        email: 'usuario@ejemplo.com',
        avatar: 'https://example.com/avatar.jpg',
        walletAddress: '0x1234567890123456789012345678901234567890',
        reputation: 85,
        createdAt: new Date().toISOString(),
        lastActive: new Date().toISOString()
      };

      res.json({
        success: true,
        data: user,
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al obtener usuario:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al obtener usuario',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async createSession(req: any, res: any): Promise<void> {
    try {
      const { worldId, userId, settings } = req.body;

      // Validar datos requeridos
      if (!worldId || !userId) {
        return res.status(400).json({
          success: false,
          error: {
            code: 'BAD_REQUEST',
            message: 'worldId y userId son requeridos',
            timestamp: new Date().toISOString()
          }
        });
      }

      // Simular creación de sesión
      const session = {
        id: Date.now().toString(),
        worldId,
        userId,
        status: 'active',
        position: { x: 0, y: 0, z: 0 },
        createdAt: new Date().toISOString(),
        lastActivity: new Date().toISOString()
      };

      res.status(201).json({
        success: true,
        data: session,
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al crear sesión:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al crear sesión',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async getTransactions(req: any, res: any): Promise<void> {
    try {
      const { userId, worldId, type } = req.query;

      // Simular transacciones
      const transactions = [
        {
          id: '1',
          type: 'purchase',
          amount: 100,
          currency: 'METAVERSE_TOKEN',
          from: 'user-1',
          to: 'user-2',
          worldId: 'world-1',
          timestamp: new Date().toISOString(),
          txHash: '0x1234567890123456789012345678901234567890123456789012345678901234'
        }
      ];

      res.json({
        success: true,
        data: {
          transactions,
          pagination: {
            page: 1,
            limit: 20,
            total: transactions.length,
            totalPages: 1
          }
        },
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al obtener transacciones:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al obtener transacciones',
          timestamp: new Date().toISOString()
        }
      });
    }
  }

  private async getAssets(req: any, res: any): Promise<void> {
    try {
      const { category, format } = req.query;

      // Simular assets
      const assets = [
        {
          id: '1',
          name: 'Edificio Futurista',
          category: 'building',
          format: 'gltf',
          url: 'https://example.com/assets/building.gltf',
          hash: '1234567890123456789012345678901234567890123456789012345678901234',
          price: 100,
          owner: 'user-1',
          createdAt: new Date().toISOString()
        }
      ];

      res.json({
        success: true,
        data: {
          assets,
          pagination: {
            page: 1,
            limit: 20,
            total: assets.length,
            totalPages: 1
          }
        },
        meta: {
          timestamp: new Date().toISOString(),
          version: '1.0.0',
          requestId: req.id
        }
      });

    } catch (error) {
      this.logger.error('Error al obtener assets:', error);
      res.status(500).json({
        success: false,
        error: {
          code: 'INTERNAL_ERROR',
          message: 'Error al obtener assets',
          timestamp: new Date().toISOString()
        }
      });
    }
  }
} 